profile_and_auth/requirements.txt
Flask
flask-restful
flask-mysql
PyJWT
Pillow
flask-cors
boto3